+++ example/lib/login_page.dart
@@ -12,20 +12,39 @@ class _LoginPageState extends State<LoginPage> {
   @override
   Widget build(BuildContext context) {
     return LoginForm(
-      callback: (email,password){
+      callback: (email, password) {
         print("$email,$password");
       },
       buttonLoginDecorationElement: ButtonLoginDecorationElement(
-        backgroundColorButton: Colors.amber,
-        widthSubmitButton: 200,
-        radiusBorderButton: 10,
-      ),
+          backgroundColorButton: Colors.white,
+          widthSubmitButton: 200,
+          shapeButtonLogin: StadiumBorder().copyWith(
+            side: BorderSide(
+              color: Colors.amber,
+              width: 0.6,
+            ),
+          ),
+          elevation: 0.0),
       onlyEmail: false,
       labelLogin: "Username",
       password: "Password",
       textButton: Text("Log IN"),
-      decorationElement: RoundedDecorationElement(
-        filledColor: Colors.grey[300],
+      paddingFields: const EdgeInsets.all(0),
+      decorationEmailElement: OutlineDecorationElement(
+        filledColor: Colors.white,
+        radius: BorderRadius.only(
+          topLeft: Radius.circular(5.0),
+          topRight: Radius.circular(5.0),
+        ),
+        widthSide: 0.6,
+      ),
+      decorationPasswordElement: OutlineDecorationElement(
+        filledColor: Colors.white,
+        radius: BorderRadius.only(
+          bottomLeft: Radius.circular(5.0),
+          bottomRight: Radius.circular(5.0),
+        ),
+        widthSide: 0.6,
       ),
     );
   }