+++ lib/dynamic_form.dart
@@ -1,8 +1,8 @@
 library dynamic_form;
 
-export 'src/simple_dynamic_form.dart';
-export 'src/login_form.dart';
+export 'src/decoration_element.dart';
 export 'src/element.dart';
 export 'src/group_elements.dart';
-export 'src/decoration_element.dart';
+export 'src/login_form.dart';
+export 'src/simple_dynamic_form.dart';
 export 'src/utilities/patterns.dart';